backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
boto3==1.36.22
botocore==1.36.22
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
Flask==3.1.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
